Cleanup singleTeam route
